Model of the Stability of Alienation

The data comes from research on the stability of alienation using panel data. The variables in the model were aggregated from 932 individual responses drawn from a rural region of Illinois to a series of items from attitude scales measured in two time points (1967 and 1971). "Anomia" and "powerlessness" were both conceived as indicators of the latent variable "alienation". Background variables including "education", and the socio-economic index score ("SEI") were also measured in 1967 and were boh conceived as indicators of the respondent's socio-economic status.

alienation

Format

A 6 by 6 variance-covariance matrix summarized from 932 respondent.

anomia67

anomia measured in 1967.

powerlessness67

powerlessness measured in 1967

anomia71

anomia measured in 1971.

powerlessness71

powerlessness measured in 1971

education

measured as years of schooling

sei

Duncan's socio-economic index score ("SEI") of the respondent's occupation

Source

From Wheaton et al. (1977).

References

Wheaton, B., Muthén, B., Alwin, D., & Summers, G. (1977). Assessing Reliability and Stability in Panel Models. Sociological Methodology, 8, 84-136. doi:10.10.2307/270754.

Examples

data(alienation, package = "jeksterslabRdatarepo")
str(alienation)

#>  num [1:6, 1:6] 11.83 6.95 6.82 4.78 -3.84 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:6] "anomia67" "powerless67" "anomia71" "powerless71" ...
#>   ..$ : chr [1:6] "anomia67" "powerless67" "anomia71" "powerless71" ...

alienation

#>             anomia67 powerless67 anomia71 powerless71 education     sei
#> anomia67      11.834       6.947    6.819       4.783    -3.839 -21.899
#> powerless67    6.947       9.364    5.091       5.028    -3.889 -18.831
#> anomia71       6.819       5.091   12.532       7.495    -3.841 -21.748
#> powerless71    4.783       5.028    7.495       9.986    -3.625 -18.775
#> education     -3.839      -3.889   -3.841      -3.625     9.610  35.522
#> sei          -21.899     -18.831  -21.748     -18.775    35.522 450.288