Galton's Height Data

This is the famous Galton data on the heights or parents and their children (i.e., where the term "regression" comes from).

galton

Format

A data frame with 898 cases and 6 variables:

family

The family that the child belongs to, labeled by the numbers from 1 to 204 and 136A.

father

The father's height, in inches.

mother

The mother's height, in inches.

gender

The gender of the child, male (M) or female (F).

height

The height of the child, in inches.

kids

The number of kids in the family of the child.

male

1 if the child is male. 0 if the child is female.

female

1 if the child is female. 0 if the child is male.

Source

Francis Galton, 2017, "Galton height data", https://doi.org/10.7910/DVN/T0HSJ1, Harvard Dataverse, V1, UNF:6:2ty+0YgqR2a66FlvjCuPkQ== [fileUNF]

References

Galton, F. (1886). Regression Towards Mediocrity in Hereditary Stature. The Journal of the Anthropological Institute of Great Britain and Ireland, 15, 246-263. doi:10.2307/2841583.

Wikipedia: Regression toward the mean

The troubling legacy of Francis Galton

Examples

data(galton, package = "jeksterslabRdatarepo")
str(galton)

#> 'data.frame':	898 obs. of  8 variables:
#>  $ family: Factor w/ 197 levels "1","10","100",..: 1 1 1 1 108 108 108 108 123 123 ...
#>  $ father: num  78.5 78.5 78.5 78.5 75.5 75.5 75.5 75.5 75 75 ...
#>  $ mother: num  67 67 67 67 66.5 66.5 66.5 66.5 64 64 ...
#>  $ gender: Factor w/ 2 levels "F","M": 2 1 1 1 2 2 1 1 2 1 ...
#>  $ height: num  73.2 69.2 69 69 73.5 72.5 65.5 65.5 71 68 ...
#>  $ kids  : int  4 4 4 4 4 4 4 4 2 2 ...
#>  $ male  : num  1 0 0 0 1 1 0 0 1 0 ...
#>  $ female: num  0 1 1 1 0 0 1 1 0 1 ...

head(galton)

#>   family father mother gender height kids male female
#> 1      1   78.5   67.0      M   73.2    4    1      0
#> 2      1   78.5   67.0      F   69.2    4    0      1
#> 3      1   78.5   67.0      F   69.0    4    0      1
#> 4      1   78.5   67.0      F   69.0    4    0      1
#> 5      2   75.5   66.5      M   73.5    4    1      0
#> 6      2   75.5   66.5      M   72.5    4    1      0