Law School Data

The population of measurements (LSAT, GPA), for the universe of 82 American law schools participating in a large study of admission practices. Two measurements were made on the entering classes of each school in 1973: LSAT, the average score for the class on a national law test, and GPA, the average undergraduate grade-point average for the class.

law

Format

A data frame with 82 cases and 2 variables:

LSAT

Law School Admission Test.

GPA

Grade Point Average.

Source

Table 3.2 in Efron and Tibshirani (1993).

References

Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York, N.Y: Chapman & Hall. ISBN:9780412042317. Google:gLlpIUxRntoC.

Examples

data(law, package = "jeksterslabRdatarepo")
str(law)

#> 'data.frame':	82 obs. of  2 variables:
#>  $ LSAT: int  622 542 579 653 606 576 620 615 553 607 ...
#>  $ GPA : num  3.23 2.83 3.24 3.12 3.09 3.39 3.1 3.4 2.97 2.91 ...

head(law)

#>   LSAT  GPA
#> 1  622 3.23
#> 2  542 2.83
#> 3  579 3.24
#> 4  653 3.12
#> 5  606 3.09
#> 6  576 3.39