Singular Value Decomposition in Regression Analysis

This data set can be used to demostrate the use of singular value decomposition in regression analysis.

svd

Format

A list with two elements y and X. y is a vector of length 8 and X is an 8 by 3 matrix.

Source

Table 1 in Mandel (1982)

References

Mandel, J. (1982). Use of the Singular Value Decomposition in Regression Analysis. The American Statistician, 36(1), 15-24. doi:10.1080/00031305.1982.10482771.

Examples

data(svd, package = "jeksterslabRdatarepo")
str(svd)

#> List of 2
#>  $ y: num [1:8] 41.4 31 37.4 50 39.2 ...
#>  $ X: num [1:8, 1:3] 1 1 1 1 1 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"

head(svd)

#> $y
#> [1] 41.38 31.01 37.41 50.05 39.17 38.86 46.14 44.47
#> 
#> $X
#>      x1    x2    x3
#> [1,]  1 16.85  1.46
#> [2,]  1 24.81 -4.61
#> [3,]  1 18.85 -0.21
#> [4,]  1 12.63  4.93
#> [5,]  1 21.38 -1.36
#> [6,]  1 18.78 -0.08
#> [7,]  1 15.58  2.98
#> [8,]  1 16.30  1.73
#>